+++ crates/amaru-ledger/src/store/columns/dreps.rs
@@ -32,7 +32,7 @@ pub struct Row {
     pub valid_until: Epoch,
 }
 
-impl<C> cbor::encode::Encode<C> for Row {
+impl<C: cbor::HasProtocolVersion> cbor::encode::Encode<C> for Row {
     fn encode<W: cbor::encode::Write>(
         &self,
         e: &mut cbor::Encoder<W>,
@@ -47,7 +47,7 @@ impl<C> cbor::encode::Encode<C> for Row {
     }
 }
 
-impl<'a, C> cbor::decode::Decode<'a, C> for Row {
+impl<'a, C: cbor::HasProtocolVersion> cbor::decode::Decode<'a, C> for Row {
     fn decode(d: &mut cbor::Decoder<'a>, ctx: &mut C) -> Result<Self, cbor::decode::Error> {
         d.array()?;
         Ok(Row {