+++ crates/amaru-ledger/src/store/columns/pools.rs
@@ -51,7 +51,7 @@ impl Row {
     }
 }
 
-impl<C> cbor::encode::Encode<C> for Row {
+impl<C: cbor::HasProtocolVersion> cbor::encode::Encode<C> for Row {
     fn encode<W: cbor::encode::Write>(
         &self,
         e: &mut cbor::Encoder<W>,
@@ -66,7 +66,7 @@ impl<C> cbor::encode::Encode<C> for Row {
     }
 }
 
-impl<'a, C> cbor::decode::Decode<'a, C> for Row {
+impl<'a, C: cbor::HasProtocolVersion> cbor::decode::Decode<'a, C> for Row {
     fn decode(d: &mut cbor::Decoder<'a>, ctx: &mut C) -> Result<Self, cbor::decode::Error> {
         d.array()?;
         let registered_at = d.decode_with(ctx)?;